adding simulation.py changes to poisson and IO
